Separate Movement Summary from Expense service
