+++ PFM.Api/src/PFM.Services/Interfaces/IExpenseService.cs
@@ -1,7 +1,4 @@
-﻿using PFM.Api.Contracts.BudgetPlan;
-using PFM.Api.Contracts.Dashboard;
-using PFM.Api.Contracts.Expense;
-using System;
+﻿using PFM.Api.Contracts.Expense;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -18,8 +15,6 @@ namespace PFM.Services.Interfaces
         Task<ExpenseDetails> GetById(int id);
 
         Task<bool> ChangeDebitStatus(int id, bool debitStatus);
-
-        Task<ExpenseSummary> GetExpenseSummary(int accountId, BudgetPlanDetails budgetPlan, DateTime referenceDate);
 
         Task<IList<ExpenseList>> GetExpenses(Api.Contracts.SearchParameters.ExpenseGetListSearchParameters search);
     }
